Migrate Dockerfiles from external repository (#7303)
* migrate OpenMS Dockerfile to main repository

* base docker images on external contrib image

* use separate build argument for OpenMS version

this is used in used for two things:
- selecting the base contrib image version
- selecting the branch from the OpenMS repository

* add stage-specific LABEL metadata

* fix stage labels, disambiguate version and base image build args, and update workflow accordingly

* add GHA cache to docker build workflow

* use more convervative default for NUM_BUILD_CORES build arg

NOTE: in github actions this is set to 2

* add a slimmer tools stage

a few other modifications:
- added dockerignore file
- clone thirdparty repo instead of using submodule since I've added .git to the dockergignore

* fix missing checkout action

* fix missing Dockerfile

* go back to cloning inside build

copying from the CWD gave me a bunch of issues while testing locally (eg. wrong context, accidental cache invalidation etc.) that I think it's better to just `git clone`.

TODO: ensure we're not unnecessarily including the .git folder now

* run `update-ca-certificates` in same command as the package installs

* remove accidental newline at beginning of dockerignore

* fix cmake_prefix_path setting

single quotes meant that we were literally adding ${CONTRIB_BUILD_DIR} to the prefix path

* make use of THIRDPARTY_DIR build arg instead of hardcoding

maybe I've gone overkill on the build args here but I think it helps reduce the chance of errors when dealing with multiple stages/images and makes it easier to move things around

* use 4 build cores for container deploy action

* remove Sirius from docker image path

Sirius adapters are in the process of being removed in #7234

* move install in docker image from /OpenMS to /opt/OpenMS

* install JRE for docker tools runtime dependency

* add a library-build stage to minimize the size of the library container

also includes some fixes to ensure that the OpenMS.git does not come along for the ride in the final runtime images

* test both tools and library runtime docker images

* actually copy in the thirdparty deps..

* ensure we fail out of any multiline scripts

* remove -o pipefail since docker RUN commands are run in /bin/sh which does not support this options

* add a contrib stage as an alternative base to that hosted in OpenMS/contrib

NOTE: this is mostly for testing and local builds and it is not the default.

* fix docker contrib stage

we need a bin subdirectory.

* move cmake install after other packages

* install coinor runtime dependency

* docker: add mono runtime dependency for ThermoRawFileParser

* docker: fix DEBIAN_FRONTEND setting

* docker: fix to apply DEBIAN_FRONTEND envvar properly

* docker: separate tools and tools+thirdparty images

* docker: switch back to copying source from context instead of using git clone

Rationale for this is to make it possible to build this in forks or locally

* docker: add readme describing images

can do more work here

* docker: add dependency for running tests

* docker: create the contrib build dir if it doesn't exist

* docker: use `make install` and refactor to use single build stage

also:
- renamed some docker ARGs.. we don't need to prefix everything with OPENMS_

* docker: tweak workflow to ensure we have a .git folder during build so we can submodule update

* docker: keep git dir when building docker images

* docker: remove gha cache scope arg

this was overly specific. we might want to have a scope eventually but I don't think caching was really working before. I'm finding that information on this is a bit sparse, perhaps this makes sense given that this is labeled "experimental"..

* docker: rename OPENMS_VERSION ARG to OPENMS_VERSION_TAG for clarity

* docker: set contrib image tag to latest unless

this is a complication of having the contrib dockerfile built in a separate repository. During a release, we build it here and the tag is consistent between the contrib image and the other images we build. For non-release pushes, we have to make sure that the tag we use to pull that image is "latest" or else this action will fail on other branches that may not exist in the contrib repository.

* docker: run tests as first step of workflow

* docker: use consistent case for "AS"

* docker: remove currently unused pyopenms stage

* docker: fix tests not passing due to missing examples

* docker: copy share from install directory now that #7337 has been merged

* docker: update readme and add changelog entry

* docker: remove unnecessary arg

* docker: whitespace differences

* docker: do not push test image

it is sufficient for caching purposes to load it

* docker: free up space on runner by deleting unneccessary tools

runners have varying amounts of free space, and these images can be quite big and sometimes juuuust fail to build if the runner happens to have less free space.

* remove contrib stage (was only used in testing)

---------

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>
Co-authored-by: Samuel Wein <sam@samwein.com>